% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newscript.R
\name{newscript}
\alias{newscript}
\title{New script generation from a template}
\usage{
newscript(filepath, title)
}
\arguments{
\item{filepath}{full path to file, including file name}

\item{title}{title for Rmd output}
}
\description{
Generates a new script with custom Rmd header from a template, updates it with your titles and current date, and opens it for editing.
}
\examples{
newscript(paste0(mypath, 'newfxn.R'),'Auto-generated R script')
}
\keyword{file}
\keyword{open}
\keyword{template}

